feat(LocationView): add available_locations action to filter active orders
Add a new action `available_locations` to the `LocationViewSet` to retrieve locations without a parent and with active orders. This enhances the API by providing a way to fetch locations that are currently in use for active orders.

refactor(Orders): add related_name to ForeignKey fields

Add `related_name='orders'` to the `restaurant` and `location` ForeignKey fields in the `Orders` model to improve querying capabilities and maintain consistency in related object access.

chore(migrations): create migration for ForeignKey related_name changes

Generate a new migration file to apply the changes made to the `Orders` model, specifically the addition of `related_name` to the `restaurant` and `location` fields.